+++ CHANGELOG.rst
@@ -14,6 +14,13 @@ Changelog
     Note that they these tags will not actually close the issue/PR until they
     are merged into the "default" branch.
 
+v0.6.7
+-------
+
+Fix:
+
+- Bug for Delete Player
+
 v0.6.6
 -------
 
@@ -245,7 +252,7 @@ v0.4.0
 
 Feature:
 
-- Item Editor with Autocomplete Combobox 
+- Item Editor with Autocomplete Combobox
 
 v0.3.10
 -------
@@ -285,7 +292,7 @@ v0.3.6
 
 Feature:
 
-- Move Guild Owner Feature 
+- Move Guild Owner Feature
 
 v0.3.4
 -------